refactor(Challenge): move out from PVZ.h
