+++ pvzclass/Classes/Challenge.cpp
@@ -1,4 +1,4 @@
-#include "..\PVZ.h"
+#include "Challenge.hpp"
 
 PVZ::Challenge::Challenge(int address)
 {
